feat: landing page 경로 설정
