+++ components/pages/landing/LandingMain.tsx
@@ -1,8 +1,15 @@
 import React from 'react';
+import { useRouter } from 'next/router';
 import * as S from './LandingMain.style';
 import mainImg from '@public/images/landing.png';
 
 function LandingMain() {
+  const router = useRouter();
+
+  const handleLoginClick = () => {
+    router.push('/login');
+  };
+
   return (
     <S.Main>
       <S.MainImg src={mainImg} alt="메인 이미지" priority />
@@ -13,7 +20,7 @@ function LandingMain() {
         </S.MainTitle>
       </S.Title>
       <S.MainSubTitle>서비스의 메인 설명 들어갑니다.</S.MainSubTitle>
-      <S.LoginButton>로그인하기</S.LoginButton>
+      <S.LoginButton onClick={handleLoginClick}>로그인하기</S.LoginButton>
     </S.Main>
   );
 }
